Split lib.rs: extract config.rs and test_support.rs
Move ServeConfig, LogFormat, AppState, ApiKey, BuiltApp, AdminConfig,
init_tracing, apply_server_defaults, and validate_serve_config into a
new config module. Move shared test fixtures (BufferWriter,
capture_json_subscriber, find_response_event, run_budget_request,
default_serve_config) into a new test_support module under #[cfg(test)].

Public API preserved via pub re-exports at the crate root. ApiKey gains
a pub new() constructor so the tuple-struct construction in build_app
works after the type moves. test_json_level_is_lowercase moves to
json_fmt.rs where it lives with the formatter it tests.
